requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
uvloop==0.19.0
httptools==0.6.1
gunicorn==22.0.0
pydantic==2.7.1
pydantic-settings==2.3.0
email-validator==2.1.1

sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

redis==5.0.4
celery[redis]==5.4.0
kombu==5.3.7
flower==2.0.1

stripe==9.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

boto3==1.34.114
botocore==1.34.114

opencv-python-headless==4.9.0.80
Pillow==10.3.0
scikit-image==0.22.0
imageio==2.34.2
imageio-ffmpeg==0.5.1

basicsr==1.4.2
realesrgan==0.3.0
facexlib==0.3.0
gfpgan==1.3.8
timm==0.9.12

librosa==0.10.2
soundfile==0.12.1
audioread==3.0.1

psutil==5.9.8
pynvml==11.5.3
sentry-sdk[fastapi]==2.5.1
python-dotenv==1.0.1
itsdangerous==2.2.0
click==8.1.7
tqdm==4.66.4
packaging==24.0
aiofiles==23.2.1
httpx==0.27.0
tenacity==8.3.0
orjson==3.10.3